# [improve][java-client] Improve performance of multi-topic consumer with more than one IO thread