--- a/pulsar-client/src/main/java/org/apache/pulsar/client/impl/MessagesImpl.java
+++ b/pulsar-client/src/main/java/org/apache/pulsar/client/impl/MessagesImpl.java
@@ -29,7 +29,7 @@ import org.apache.pulsar.client.api.Messages;
 @NotThreadSafe
 public class MessagesImpl<T> implements Messages<T> {
 
-    private List<Message<T>> messageList;
+    private final List<Message<T>> messageList;
 
     private final int maxNumberOfMessages;
     private final long maxSizeOfMessages;
@@ -78,6 +78,10 @@ public class MessagesImpl<T> implements Messages<T> {
         this.currentNumberOfMessages = 0;
         this.currentSizeOfMessages = 0;
         this.messageList.clear();
+    }
+
+    public List<Message<T>> getMessageList() {
+        return messageList;
     }
 
     @Override