--- a/pulsar-client/src/main/java/org/apache/pulsar/client/impl/MessagesImpl.java
+++ b/pulsar-client/src/main/java/org/apache/pulsar/client/impl/MessagesImpl.java
@@ -80,7 +80,7 @@ public class MessagesImpl<T> implements Messages<T> {
         this.messageList.clear();
     }
 
-    public List<Message<T>> getMessageList() {
+    List<Message<T>> getMessageList() {
         return messageList;
     }